Добавление модульности. Разбиение кода по файлам
