+++ js/data.js
@@ -1,0 +1,43 @@
+function getMessages() {
+  const MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+  return MESSAGES;
+}
+
+function getNames() {
+  const NAMES = [
+    'Мирослава',
+    'Александр',
+    'Иван',
+    'Федор',
+    'Сергей',
+    'Алексей',
+    'Светлана',
+    'Евгения',
+  ];
+
+  return NAMES;
+}
+
+function getDescription() {
+  const DESCRIPTIONS = [
+    'На море!',
+    'Зарядитесь нашим теплом.',
+    'Просто я',
+    'Мыслями на пляже.',
+    'Я после марафона… любимого сериала.',
+    'Работаю как босс',
+    'Старые подходы не помогут открыть новые двери'
+  ];
+
+  return DESCRIPTIONS;
+}
+
+export { getMessages, getNames, getDescription };
